fix: explicitly cast both tl.dot operands to float32
tl.load other=0.0 upcast behavior varies by Triton version. Cast both
a and b to float32 before tl.dot to guarantee matching dtypes everywhere.

+++ rocm_qlora/kernels/dequant_matmul.py
@@ -89,9 +89,8 @@ def dequant_int8_matmul_kernel(
         # Dequantize B
         b_fp32 = b_int8.to(tl.float32) * scales.to(tl.float32)
         
-        # Matrix multiply (tl.dot expects [M, K] and [K, N])
-        # b_fp32 is [BLOCK_K, BLOCK_N] (since b_ptrs was [BLOCK_K, BLOCK_N])
-        acc += tl.dot(a, b_fp32)
+        # Matrix multiply — both operands must be same dtype for tl.dot.
+        acc += tl.dot(a.to(tl.float32), b_fp32)
         
         # Advance pointers
         a_ptrs += BLOCK_K * stride_ak
@@ -184,7 +183,7 @@ def dequant_nf4_matmul_kernel(
         
         b_fp32 = b_val.to(tl.float32) * scales.to(tl.float32)
         
-        acc += tl.dot(a, b_fp32)
+        acc += tl.dot(a.to(tl.float32), b_fp32)
         
         a_ptrs += BLOCK_K * stride_ak
